pyproject.toml
[tool.poetry]
name = "otel-datadog-tests"
version = "0.1.0"
description = "Just a test repo to verify OpenTelemetry with Datadog"
authors = ["Diogo Baeder <diogobaeder@yahoo.com.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
gunicorn = "^23.0.0"
json-log-formatter = "^1.0"
celery = "^5.4.0"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
gevent = "^24.2.1"
opentelemetry-distro = "^0.48b0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-logging = "^0.48b0"
opentelemetry-instrumentation-wsgi = "^0.48b0"
opentelemetry-instrumentation-celery = "^0.48b0"
opentelemetry-instrumentation-flask = "^0.48b0"
opentelemetry-instrumentation-grpc = "^0.48b0"
opentelemetry-instrumentation-requests = "^0.48b0"
opentelemetry-instrumentation-sqlalchemy = "^0.48b0"
opentelemetry-instrumentation-urllib3 = "^0.48b0"
opentelemetry-instrumentation-urllib = "^0.48b0"
opentelemetry-instrumentation-threading = "^0.48b0"
opentelemetry-instrumentation-dbapi = "^0.48b0"
redis = "^5.0.8"
opentelemetry-instrumentation-redis = "^0.48b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
